+++ libs/core/src/lib/wizard/wizard.module.ts
@@ -8,6 +8,7 @@ import { WizardStepIndicatorComponent } from './wizard-step-indicator/wizard-ste
 import { WizardContentComponent } from './wizard-content/wizard-content.component';
 import { WizardNextStepComponent } from './wizard-next-step/wizard-next-step.component';
 import { IconModule } from '../icon/icon.module';
+import { ActionSheetModule } from '../action-sheet/action-sheet.module';
 import { ScrollSpyModule } from '../scroll-spy/scroll-spy.module';
 
 @NgModule({
@@ -20,7 +21,7 @@ import { ScrollSpyModule } from '../scroll-spy/scroll-spy.module';
         WizardContentComponent,
         WizardNextStepComponent
     ],
-    imports: [CommonModule, IconModule, ScrollSpyModule],
+    imports: [CommonModule, IconModule, ScrollSpyModule, ActionSheetModule],
     exports: [
         WizardComponent,
         WizardNavigationComponent,